Remove card page, comments now inline modal like the calendar
